refactor(#808): extract import collection/registration from index.ts → registry/imports.ts
God-file breakdown pilot. Moves the 9 import-infrastructure functions
(collectAllSourceImports, addStringImports, addUnionImports,
addUnionImportsAsNativeFuncs, addIteratorImports, addArrayIteratorImports,
addGeneratorImports, addForInImports, collectUsedExternImports — ~1,789 LOC)
out of the src/codegen/index.ts god-file (15,749 → 13,951) into the cohesive
src/codegen/registry/imports.ts (417 → 2,237, where addImport already lives).

Behaviour-preserving VERBATIM move, byte-identity guarded:
- npx tsx scripts/prove-emit-identity.mjs check → IDENTICAL (39/39 file,target,
  gc/standalone/wasi) — the issue's 'same .wasm output' acceptance criterion.
- tsc --noEmit = 0. Smoke: gc + standalone both compile+run (Math/Array/typeof-any
  union path, ret=23), exercising the addUnionImportsAsNativeFuncs native synthesis.

Wiring: index.ts re-exports the 6 externally-imported functions (so every
existing `from "./index.js"` importer keeps resolving) and imports back the 4
it calls internally; imports.ts imports the moved code's deps (rebased paths)
plus isTupleType + isStandaloneRegExpMatchArrayValue (now exported from index.ts).
The addUnionImports index-shift invariant is preserved by construction (verbatim).

loc-budget-allow granted for the consolidated registry/imports.ts (#3131).

✓
